+++ backend/src/test/java/hanglog/global/ControllerTest.java
@@ -5,7 +5,7 @@ import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.docu
 import hanglog.auth.AuthArgumentResolver;
 import hanglog.auth.domain.BearerAuthorizationExtractor;
 import hanglog.auth.domain.JwtProvider;
-import hanglog.trip.restdocs.RestDocsConfiguration;
+import hanglog.global.restdocs.RestDocsConfiguration;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
