+++ backend/src/test/java/hanglog/global/restdocs/RestDocsConfiguration.java
@@ -1,4 +1,4 @@
-package hanglog.trip.restdocs;
+package hanglog.global.restdocs;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
