+++ backend/src/test/java/hanglog/integration/controller/SharedTripIntegrationTest.java
@@ -1,19 +1,17 @@
-package hanglog.share.integration;
+package hanglog.integration.controller;
 
-import static hanglog.global.IntegrationFixture.END_DATE;
-import static hanglog.global.IntegrationFixture.START_DATE;
 import static hanglog.global.exception.ExceptionCode.INVALID_SHARE_CODE;
 import static hanglog.global.exception.ExceptionCode.NOT_FOUND_SHARED_CODE;
+import static hanglog.integration.IntegrationFixture.END_DATE;
+import static hanglog.integration.IntegrationFixture.START_DATE;
 import static io.restassured.http.ContentType.JSON;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 
 import hanglog.auth.domain.MemberTokens;
-import hanglog.global.IntegrationTest;
 import hanglog.share.dto.request.SharedTripStatusRequest;
 import hanglog.trip.dto.request.TripCreateRequest;
-import hanglog.trip.integration.TripIntegrationTest;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
@@ -82,9 +80,9 @@ class SharedTripIntegrationTest extends IntegrationTest {
 
     @DisplayName("처음 생성된 여행은 비공유 상태이다")
     @Test
-    void updateSharedStatus_InitialTripStatus(){
+    void updateSharedStatus_InitialTripStatus() {
         // when
-        final ExtractableResponse<Response> response = requestGetTrip(memberTokens,tripId);
+        final ExtractableResponse<Response> response = requestGetTrip(memberTokens, tripId);
 
         // then
         assertThat(response.body().jsonPath().getString("sharedTrip")).isNull();
@@ -92,13 +90,13 @@ class SharedTripIntegrationTest extends IntegrationTest {
 
     @DisplayName("비공유 여행에서는 공유 코드를 볼 수 없다")
     @Test
-    void updateSharedStatus_UnSharedTrip(){
+    void updateSharedStatus_UnSharedTrip() {
         // given
         requestUpdateSharedTripStatus(true);
         requestUpdateSharedTripStatus(false);
 
         // when
-        final ExtractableResponse<Response> response = requestGetTrip(memberTokens,tripId);
+        final ExtractableResponse<Response> response = requestGetTrip(memberTokens, tripId);
 
         // then
         assertThat(response.body().jsonPath().getString("sharedTrip")).isNull();
@@ -106,10 +104,10 @@ class SharedTripIntegrationTest extends IntegrationTest {
 
     @DisplayName("삭제된 여행은 공유 할 수 없다")
     @Test
-    void getSharedTrip_deleteTripFail(){
+    void getSharedTrip_deleteTripFail() {
         // given
         final String sharedCode = requestUpdateSharedTripStatus(true).body().jsonPath().get("sharedCode");
-        requestDeleteTrip(memberTokens,tripId);
+        requestDeleteTrip(memberTokens, tripId);
 
         // when
         final ExtractableResponse<Response> response = RestAssured.given()
