+++ backend/src/test/java/hanglog/integration/service/ServiceIntegrationTest.java
@@ -1,4 +1,4 @@
-package hanglog.global;
+package hanglog.integration.service;
 
 import hanglog.member.domain.Member;
 import hanglog.member.domain.repository.MemberRepository;
