+++ backend/src/test/java/hanglog/trip/presentation/ItemControllerTest.java
@@ -1,6 +1,6 @@
 package hanglog.trip.presentation;
 
-import static hanglog.trip.restdocs.RestDocsConfiguration.field;
+import static hanglog.global.restdocs.RestDocsConfiguration.field;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
